moved the calls back
I moved the calls for lidHookOutside();   buildLightTubes(); buildButtons(); lidHookInside(); back to the original locations.  This fixed the button issue (and I'm assuming the inside/outside issues).  I don't have any code for the inside/outside modules to test with.

But it does break (well maybe not break but change) the inspectx/y functionality. But I do see that you had separate inspect options for the light tubes and buttons so having them as part of the inspect function wasn't there before.  If I get a chance I may look at how to get the inspectx/y working for everything but as you have the other options it's probably not necessary.