Code Cleanup and PCB Thickness fix
While setting up another box I discovered that I needed to have a gap in the stands and connectors different than the default.  I had previously changed the names of the standoffXXX parameters to defaultStandoffXXX and that messed with some of the examples you created.

Added optional PCB Gap to the Standoffs and Connectors.  Both are optional and don't change the results if they are omitted.

I also printed a box and discovered that the diamond snaps didn't line up correctly.  That was fixed and I adjusted the size of all snaps to be calculated from the wall thickness.

I also cleaned dead code and fixed some indentation/spacing issues.