Hopefully final update for 3.0 release.
Fixed all issues reported.

Updated test files and added a few more.

Changed how Masks are defined and used.  Should be simpler now.  Just add the MaskName or optionally add a vector with the mask name H Offset, V Offset and Rotation.  for example [maskHoneycomb, 0,2,30].

Added the option of negative depth text (to make it stand out of the plane instead of going in.  Also limited the depth so that you won't cut holes in the case.

Added a Z inspection option.

Changed how the mask is applied to only generate the minimal size to cover the opening.

Removed the Post Hook names so they are either normal or Pre.

Fixed Snap Joint positions for all of the allowed coordinate options (and added test file)

Added Sanity checks to all objects.

Added test modules for Masks and Polygons.