Change Coord defaults
also added using the 2 coordinate systems in the Ridge Extensions.  Since it was at an odd reference before (the outside of the shell for horizontal and the ridge height for the vertical origin I changed how it works.

So now if you have PCB coordinates it starts at the edge of the PCB horizontally and the top of the PCB vertically with negative being below the PCB.  In Box it goes from the origin with only positive heights.