redo of buttons
When I went to add some value checking to the buttons I saw that the code had different versions for the preview and so I merged it and added value checking.

I also added optional parameters for the predefined buttonWall, buttonPlateThickness and buttonSlack.  The cup depth is now calculated based on the other variables.