Added undef alternates
Looking at the screenshot the artifact looks like it could be the "hidden" surfaces of the box used to trim off the half of the base that isn't used.  I added a union to try and get around a OpenSCAD issue between the different preview systems on different OS's.  Only found a few references to there being some slight differences in preview mode, but render mode should be consistent.

Let me know if that resolves it on your Mac.